backend/requirements.txt
asgiref==3.5.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
Deprecated==1.2.13
Django==4.0.4
django-cors-headers==3.11.0
djangorestframework==3.13.1
finnhub-python==2.4.13
idna==3.3
Jinja2==3.1.1
joblib==1.1.0
MarkupSafe==2.1.1
nltk==3.7
numpy==1.22.3
packaging==21.3
pandas==1.4.2
praw==7.5.0
prawcore==2.3.0
psycopg2==2.9.3
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
redis==4.2.2
regex==2022.3.15
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tqdm==4.64.0
update-checker==0.18.0
urllib3==1.26.9
websocket-client==1.3.2
wrapt==1.14.0
